Rename server_nonce to server_nonce_var contextvar
