+++ mtpylon/service_schema/functions/__init__.py
@@ -2,7 +2,7 @@
 from .req_pq_func import req_pq
 from .req_pq_multi_func import req_pq_multi
 from .req_DH_params_func import req_DH_params
-from .set_client_DH_params import set_client_DH_params
+from .set_client_DH_params_func import set_client_DH_params
 from .rpc_drop_answer import rpc_drop_answer
 from .get_future_salts import get_future_salts
 from .ping import ping
