+++ tests/test_service_schema/test_functions/test_req_pq_func.py
@@ -6,7 +6,7 @@ import pytest
 
 from mtpylon import int128
 from mtpylon.contextvars import (
-    server_nonce,
+    server_nonce_var,
     p_var,
     q_var,
     pq_var,
@@ -49,7 +49,7 @@ async def test_req_pq():
 
         assert result.nonce == nonce_value
         assert result.server_nonce == server_nonce_value
-        assert server_nonce.get() == server_nonce_value
+        assert server_nonce_var.get() == server_nonce_value
 
         assert int.from_bytes(result.pq, 'big') == pq_value
         assert pq_var.get() == pq_value
